+++ src/layout.cpp
@@ -1460,7 +1460,7 @@ void LayoutParser::endElement( const std::string &name )
   //printf("endElement [%s]::[%s]\n",m_scope.data(),name.data());
   auto it=g_elementHandlers.end();
 
-  if (!m_scope.isEmpty() && m_scope.right(name.length()+1)==name+"/")
+  if (!m_scope.isEmpty() && m_scope.right(static_cast<uint>(name.length())+1)==name+"/")
   { // element ends current scope
     it = g_elementHandlers.find(m_scope.left(m_scope.length()-1).str());
   }
